Move sample, mode, pipe, bfill and more
