+++ python/cudf/cudf/tests/test_dataframe.py
@@ -3,7 +3,6 @@
 import array as arr
 import io
 import operator
-import re
 import textwrap
 import warnings
 from contextlib import contextmanager
@@ -49,32 +48,6 @@ if get_global_manager() is not None:
     NUMERIC_TYPES = ["float32"]
     # To save time, we skip tests marked "xfail"
     pytest_xfail = pytest.mark.skipif
-
-
-@contextmanager
-def _hide_ufunc_warnings(eval_str):
-    # pandas raises warnings for some inputs to the following ufuncs:
-    if any(
-        x in eval_str
-        for x in {
-            "arctanh",
-            "log",
-        }
-    ):
-        with warnings.catch_warnings():
-            warnings.filterwarnings(
-                "ignore",
-                "invalid value encountered in",
-                category=RuntimeWarning,
-            )
-            warnings.filterwarnings(
-                "ignore",
-                "divide by zero encountered in",
-                category=RuntimeWarning,
-            )
-            yield
-    else:
-        yield
 
 
 @contextmanager
@@ -3605,42 +3578,6 @@ def test_dataframe_concat_dataframe_lists(df, other, sort, ignore_index):
         )
 
 
-@pytest.mark.parametrize(
-    "df",
-    [
-        pd.DataFrame({"A": [1, 2, 3, np.nan, None, 6]}),
-        pd.Series([1, 2, 3, None, np.nan, 5, 6, np.nan]),
-    ],
-)
-@pytest.mark.parametrize("alias", ["bfill", "backfill"])
-def test_dataframe_bfill(df, alias):
-    gdf = cudf.from_pandas(df)
-
-    with expect_warning_if(alias == "backfill"):
-        actual = getattr(df, alias)()
-    with expect_warning_if(alias == "backfill"):
-        expected = getattr(gdf, alias)()
-    assert_eq(expected, actual)
-
-
-@pytest.mark.parametrize(
-    "df",
-    [
-        pd.DataFrame({"A": [1, 2, 3, np.nan, None, 6]}),
-        pd.Series([1, 2, 3, None, np.nan, 5, 6, np.nan]),
-    ],
-)
-@pytest.mark.parametrize("alias", ["ffill", "pad"])
-def test_dataframe_ffill(df, alias):
-    gdf = cudf.from_pandas(df)
-
-    with expect_warning_if(alias == "pad"):
-        actual = getattr(df, alias)()
-    with expect_warning_if(alias == "pad"):
-        expected = getattr(gdf, alias)()
-    assert_eq(expected, actual)
-
-
 @pytest_unmark_spilling
 @pytest.mark.parametrize(
     "df",
@@ -3722,962 +3659,3 @@ def test_dataframe_concat_series_without_name():
     ps = gs.to_pandas()
 
     assert_eq(pd.concat([pdf, ps]), cudf.concat([df, gs]))
-
-
-def test_cudf_arrow_array_error():
-    df = cudf.DataFrame({"a": [1, 2, 3]})
-
-    with pytest.raises(
-        TypeError,
-        match="Implicit conversion to a host PyArrow object via "
-        "__arrow_array__ is not allowed. Consider using .to_arrow()",
-    ):
-        df.__arrow_array__()
-
-    sr = cudf.Series([1, 2, 3])
-
-    with pytest.raises(
-        TypeError,
-        match="Implicit conversion to a host PyArrow object via "
-        "__arrow_array__ is not allowed. Consider using .to_arrow()",
-    ):
-        sr.__arrow_array__()
-
-    sr = cudf.Series(["a", "b", "c"])
-    with pytest.raises(
-        TypeError,
-        match="Implicit conversion to a host PyArrow object via "
-        "__arrow_array__ is not allowed. Consider using .to_arrow()",
-    ):
-        sr.__arrow_array__()
-
-
-@pytest.mark.parametrize(
-    "make_weights_axis_1",
-    [lambda _: None, lambda s: [1] * s, lambda s: np.ones(s)],
-)
-def test_sample_axis_1(
-    sample_n_frac, random_state_tuple_axis_1, make_weights_axis_1
-):
-    n, frac = sample_n_frac
-    pd_random_state, gd_random_state, checker = random_state_tuple_axis_1
-
-    pdf = pd.DataFrame(
-        {
-            "a": [1, 2, 3, 4, 5],
-            "float": [0.05, 0.2, 0.3, 0.2, 0.25],
-            "int": [1, 3, 5, 4, 2],
-        },
-    )
-    df = cudf.DataFrame.from_pandas(pdf)
-
-    weights = make_weights_axis_1(len(pdf.columns))
-
-    expected = pdf.sample(
-        n=n,
-        frac=frac,
-        replace=False,
-        random_state=pd_random_state,
-        weights=weights,
-        axis=1,
-    )
-    got = df.sample(
-        n=n,
-        frac=frac,
-        replace=False,
-        random_state=gd_random_state,
-        weights=weights,
-        axis=1,
-    )
-    checker(expected, got)
-
-
-@pytest.mark.parametrize(
-    "pdf",
-    [
-        pd.DataFrame(
-            {
-                "a": [1, 2, 3, 4, 5],
-                "float": [0.05, 0.2, 0.3, 0.2, 0.25],
-                "int": [1, 3, 5, 4, 2],
-            },
-        ),
-        pd.Series([1, 2, 3, 4, 5]),
-    ],
-)
-@pytest.mark.parametrize("replace", [True, False])
-def test_sample_axis_0(
-    pdf, sample_n_frac, replace, random_state_tuple_axis_0, make_weights_axis_0
-):
-    n, frac = sample_n_frac
-    pd_random_state, gd_random_state, checker = random_state_tuple_axis_0
-
-    df = cudf.from_pandas(pdf)
-
-    pd_weights, gd_weights = make_weights_axis_0(
-        len(pdf), isinstance(gd_random_state, np.random.RandomState)
-    )
-    if (
-        not replace
-        and not isinstance(gd_random_state, np.random.RandomState)
-        and gd_weights is not None
-    ):
-        pytest.skip(
-            "`cupy.random.RandomState` doesn't support weighted sampling "
-            "without replacement."
-        )
-
-    expected = pdf.sample(
-        n=n,
-        frac=frac,
-        replace=replace,
-        random_state=pd_random_state,
-        weights=pd_weights,
-        axis=0,
-    )
-
-    got = df.sample(
-        n=n,
-        frac=frac,
-        replace=replace,
-        random_state=gd_random_state,
-        weights=gd_weights,
-        axis=0,
-    )
-    checker(expected, got)
-
-
-@pytest.mark.parametrize("replace", [True, False])
-@pytest.mark.parametrize(
-    "random_state_lib", [cupy.random.RandomState, np.random.RandomState]
-)
-def test_sample_reproducibility(replace, random_state_lib):
-    df = cudf.DataFrame({"a": cupy.arange(0, 1024)})
-
-    n = 1024
-    expected = df.sample(n, replace=replace, random_state=random_state_lib(10))
-    out = df.sample(n, replace=replace, random_state=random_state_lib(10))
-
-    assert_eq(expected, out)
-
-
-@pytest.mark.parametrize("axis", [0, 1])
-def test_sample_invalid_n_frac_combo(axis):
-    n, frac = 2, 0.5
-    pdf = pd.DataFrame(
-        {
-            "a": [1, 2, 3, 4, 5],
-            "float": [0.05, 0.2, 0.3, 0.2, 0.25],
-            "int": [1, 3, 5, 4, 2],
-        },
-    )
-    df = cudf.DataFrame.from_pandas(pdf)
-
-    assert_exceptions_equal(
-        lfunc=pdf.sample,
-        rfunc=df.sample,
-        lfunc_args_and_kwargs=([], {"n": n, "frac": frac, "axis": axis}),
-        rfunc_args_and_kwargs=([], {"n": n, "frac": frac, "axis": axis}),
-    )
-
-
-@pytest.mark.parametrize("n, frac", [(100, None), (None, 3)])
-@pytest.mark.parametrize("axis", [0, 1])
-def test_oversample_without_replace(n, frac, axis):
-    pdf = pd.DataFrame({"a": [1, 2, 3, 4, 5]})
-    df = cudf.DataFrame.from_pandas(pdf)
-
-    assert_exceptions_equal(
-        lfunc=pdf.sample,
-        rfunc=df.sample,
-        lfunc_args_and_kwargs=(
-            [],
-            {"n": n, "frac": frac, "axis": axis, "replace": False},
-        ),
-        rfunc_args_and_kwargs=(
-            [],
-            {"n": n, "frac": frac, "axis": axis, "replace": False},
-        ),
-    )
-
-
-@pytest.mark.parametrize("random_state", [None, cupy.random.RandomState(42)])
-def test_sample_unsupported_arguments(random_state):
-    df = cudf.DataFrame({"float": [0.05, 0.2, 0.3, 0.2, 0.25]})
-    with pytest.raises(
-        NotImplementedError,
-        match="Random sampling with cupy does not support these inputs.",
-    ):
-        df.sample(
-            n=2, replace=False, random_state=random_state, weights=[1] * 5
-        )
-
-
-@pytest.mark.parametrize(
-    "df",
-    [
-        pd.DataFrame(),
-        pd.DataFrame(index=[100, 10, 1, 0]),
-        pd.DataFrame(columns=["a", "b", "c", "d"]),
-        pd.DataFrame(columns=["a", "b", "c", "d"], index=[100]),
-        pd.DataFrame(
-            columns=["a", "b", "c", "d"], index=[100, 10000, 2131, 133]
-        ),
-        pd.DataFrame({"a": [1, 2, 3], "b": ["abc", "xyz", "klm"]}),
-    ],
-)
-def test_dataframe_empty(df):
-    pdf = df
-    gdf = cudf.from_pandas(pdf)
-
-    assert_eq(pdf.empty, gdf.empty)
-
-
-@pytest.mark.parametrize(
-    "df",
-    [
-        pd.DataFrame(),
-        pd.DataFrame(index=[100, 10, 1, 0]),
-        pd.DataFrame(columns=["a", "b", "c", "d"]),
-        pd.DataFrame(columns=["a", "b", "c", "d"], index=[100]),
-        pd.DataFrame(
-            columns=["a", "b", "c", "d"], index=[100, 10000, 2131, 133]
-        ),
-        pd.DataFrame({"a": [1, 2, 3], "b": ["abc", "xyz", "klm"]}),
-    ],
-)
-def test_dataframe_size(df):
-    pdf = df
-    gdf = cudf.from_pandas(pdf)
-
-    assert_eq(pdf.size, gdf.size)
-
-
-@pytest.mark.parametrize(
-    "ps",
-    [
-        pd.Series(dtype="float64"),
-        pd.Series(index=[100, 10, 1, 0], dtype="float64"),
-        pd.Series([], dtype="float64"),
-        pd.Series(["a", "b", "c", "d"]),
-        pd.Series(["a", "b", "c", "d"], index=[0, 1, 10, 11]),
-    ],
-)
-def test_series_empty(ps):
-    ps = ps
-    gs = cudf.from_pandas(ps)
-
-    assert_eq(ps.empty, gs.empty)
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        None,
-        [],
-        [1],
-        {"a": [10, 11, 12]},
-        {
-            "a": [10, 11, 12],
-            "another column name": [12, 22, 34],
-            "xyz": [0, 10, 11],
-        },
-    ],
-)
-@pytest.mark.parametrize(
-    "columns",
-    [["a"], ["another column name"], None, pd.Index(["a"], name="index name")],
-)
-def test_dataframe_init_with_columns(data, columns):
-    pdf = pd.DataFrame(data, columns=columns)
-    gdf = cudf.DataFrame(data, columns=columns)
-
-    assert_eq(
-        pdf,
-        gdf,
-        check_index_type=len(pdf.index) != 0,
-        check_dtype=not (pdf.empty and len(pdf.columns)),
-        check_column_type=False,
-    )
-
-
-@pytest_unmark_spilling
-@pytest.mark.parametrize(
-    "data, ignore_dtype",
-    [
-        ([pd.Series([1, 2, 3])], False),
-        ([pd.Series(index=[1, 2, 3], dtype="float64")], False),
-        ([pd.Series(name="empty series name", dtype="float64")], False),
-        (
-            [pd.Series([1]), pd.Series([], dtype="float64"), pd.Series([3])],
-            False,
-        ),
-        (
-            [
-                pd.Series([1, 0.324234, 32424.323, -1233, 34242]),
-                pd.Series([], dtype="float64"),
-                pd.Series([3], name="series that is named"),
-            ],
-            False,
-        ),
-        ([pd.Series([1, 2, 3], name="hi")] * 10, False),
-        ([pd.Series([1, 2, 3], name=None, index=[10, 11, 12])] * 10, False),
-        (
-            [
-                pd.Series([1, 2, 3], name=None, index=[10, 11, 12]),
-                pd.Series([1, 2, 30], name=None, index=[13, 144, 15]),
-            ],
-            True,
-        ),
-        (
-            [
-                pd.Series([1, 0.324234, 32424.323, -1233, 34242]),
-                pd.Series([], dtype="float64"),
-                pd.Series(index=[10, 11, 12], dtype="float64"),
-            ],
-            False,
-        ),
-        (
-            [
-                pd.Series([1, 0.324234, 32424.323, -1233, 34242]),
-                pd.Series([], name="abc", dtype="float64"),
-                pd.Series(index=[10, 11, 12], dtype="float64"),
-            ],
-            False,
-        ),
-        (
-            [
-                pd.Series([1, 0.324234, 32424.323, -1233, 34242]),
-                pd.Series([1, -100, 200, -399, 400], name="abc"),
-                pd.Series([111, 222, 333], index=[10, 11, 12]),
-            ],
-            False,
-        ),
-    ],
-)
-@pytest.mark.parametrize(
-    "columns",
-    [
-        None,
-        ["0"],
-        [0],
-        ["abc"],
-        [144, 13],
-        [2, 1, 0],
-        pd.Index(["abc"], name="custom_name"),
-    ],
-)
-def test_dataframe_init_from_series_list(data, ignore_dtype, columns):
-    gd_data = [cudf.from_pandas(obj) for obj in data]
-
-    expected = pd.DataFrame(data, columns=columns)
-    actual = cudf.DataFrame(gd_data, columns=columns)
-
-    if ignore_dtype:
-        # When a union is performed to generate columns,
-        # the order is never guaranteed. Hence sort by
-        # columns before comparison.
-        if not expected.columns.equals(actual.columns):
-            expected = expected.sort_index(axis=1)
-            actual = actual.sort_index(axis=1)
-        assert_eq(
-            expected.fillna(-1),
-            actual.fillna(-1),
-            check_dtype=False,
-            check_index_type=True,
-        )
-    else:
-        assert_eq(
-            expected,
-            actual,
-            check_index_type=True,
-            check_column_type=False,
-        )
-
-
-@pytest_unmark_spilling
-@pytest.mark.parametrize(
-    "data, ignore_dtype, index",
-    [
-        ([pd.Series([1, 2, 3])], False, ["a", "b", "c"]),
-        ([pd.Series(index=[1, 2, 3], dtype="float64")], False, ["a", "b"]),
-        (
-            [pd.Series(name="empty series name", dtype="float64")],
-            False,
-            ["index1"],
-        ),
-        (
-            [pd.Series([1]), pd.Series([], dtype="float64"), pd.Series([3])],
-            False,
-            ["0", "2", "1"],
-        ),
-        (
-            [
-                pd.Series([1, 0.324234, 32424.323, -1233, 34242]),
-                pd.Series([], dtype="float64"),
-                pd.Series([3], name="series that is named"),
-            ],
-            False,
-            ["_", "+", "*"],
-        ),
-        ([pd.Series([1, 2, 3], name="hi")] * 10, False, ["mean"] * 10),
-        (
-            [pd.Series([1, 2, 3], name=None, index=[10, 11, 12])] * 10,
-            False,
-            ["abc"] * 10,
-        ),
-        (
-            [
-                pd.Series([1, 2, 3], name=None, index=[10, 11, 12]),
-                pd.Series([1, 2, 30], name=None, index=[13, 144, 15]),
-            ],
-            True,
-            ["set_index_a", "set_index_b"],
-        ),
-        (
-            [
-                pd.Series([1, 0.324234, 32424.323, -1233, 34242]),
-                pd.Series([], dtype="float64"),
-                pd.Series(index=[10, 11, 12], dtype="float64"),
-            ],
-            False,
-            ["a", "b", "c"],
-        ),
-        (
-            [
-                pd.Series([1, 0.324234, 32424.323, -1233, 34242]),
-                pd.Series([], name="abc", dtype="float64"),
-                pd.Series(index=[10, 11, 12], dtype="float64"),
-            ],
-            False,
-            ["a", "v", "z"],
-        ),
-        (
-            [
-                pd.Series([1, 0.324234, 32424.323, -1233, 34242]),
-                pd.Series([1, -100, 200, -399, 400], name="abc"),
-                pd.Series([111, 222, 333], index=[10, 11, 12]),
-            ],
-            False,
-            ["a", "v", "z"],
-        ),
-    ],
-)
-@pytest.mark.parametrize(
-    "columns", [None, ["0"], [0], ["abc"], [144, 13], [2, 1, 0]]
-)
-def test_dataframe_init_from_series_list_with_index(
-    data,
-    ignore_dtype,
-    index,
-    columns,
-):
-    gd_data = [cudf.from_pandas(obj) for obj in data]
-
-    expected = pd.DataFrame(data, columns=columns, index=index)
-    actual = cudf.DataFrame(gd_data, columns=columns, index=index)
-
-    if ignore_dtype:
-        # When a union is performed to generate columns,
-        # the order is never guaranteed. Hence sort by
-        # columns before comparison.
-        if not expected.columns.equals(actual.columns):
-            expected = expected.sort_index(axis=1)
-            actual = actual.sort_index(axis=1)
-        assert_eq(expected.fillna(-1), actual.fillna(-1), check_dtype=False)
-    else:
-        assert_eq(expected, actual, check_column_type=False)
-
-
-@pytest.mark.parametrize(
-    "data, index",
-    [
-        ([pd.Series([1, 2]), pd.Series([1, 2])], ["a", "b", "c"]),
-        (
-            [
-                pd.Series([1, 0.324234, 32424.323, -1233, 34242]),
-                pd.Series([], dtype="float64"),
-                pd.Series([3], name="series that is named"),
-            ],
-            ["_", "+"],
-        ),
-        ([pd.Series([1, 2, 3], name="hi")] * 10, ["mean"] * 9),
-    ],
-)
-def test_dataframe_init_from_series_list_with_index_error(data, index):
-    gd_data = [cudf.from_pandas(obj) for obj in data]
-
-    assert_exceptions_equal(
-        pd.DataFrame,
-        cudf.DataFrame,
-        ([data], {"index": index}),
-        ([gd_data], {"index": index}),
-    )
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        [pd.Series([1, 2, 3], index=["a", "a", "a"])],
-        [pd.Series([1, 2, 3], index=["a", "a", "a"])] * 4,
-        [
-            pd.Series([1, 2, 3], index=["a", "b", "a"]),
-            pd.Series([1, 2, 3], index=["b", "b", "a"]),
-        ],
-        [
-            pd.Series([1, 2, 3], index=["a", "b", "z"]),
-            pd.Series([1, 2, 3], index=["u", "b", "a"]),
-            pd.Series([1, 2, 3], index=["u", "b", "u"]),
-        ],
-    ],
-)
-def test_dataframe_init_from_series_list_duplicate_index_error(data):
-    gd_data = [cudf.from_pandas(obj) for obj in data]
-
-    assert_exceptions_equal(
-        lfunc=pd.DataFrame,
-        rfunc=cudf.DataFrame,
-        lfunc_args_and_kwargs=([], {"data": data}),
-        rfunc_args_and_kwargs=([], {"data": gd_data}),
-        check_exception_type=False,
-    )
-
-
-def test_dataframe_iterrows_itertuples():
-    df = cudf.DataFrame({"a": [1, 2, 3], "b": [4, 5, 6]})
-
-    with pytest.raises(
-        TypeError,
-        match=re.escape(
-            "cuDF does not support iteration of DataFrame "
-            "via itertuples. Consider using "
-            "`.to_pandas().itertuples()` "
-            "if you wish to iterate over namedtuples."
-        ),
-    ):
-        df.itertuples()
-
-    with pytest.raises(
-        TypeError,
-        match=re.escape(
-            "cuDF does not support iteration of DataFrame "
-            "via iterrows. Consider using "
-            "`.to_pandas().iterrows()` "
-            "if you wish to iterate over each row."
-        ),
-    ):
-        df.iterrows()
-
-
-@pytest.mark.parametrize(
-    "df",
-    [
-        lambda: cudf.DataFrame({"a": [1, 2, 3]}),
-        lambda: cudf.DataFrame(
-            {"a": [1, 2, 3], "b": ["a", "z", "c"]}, index=["a", "z", "x"]
-        ),
-        lambda: cudf.DataFrame(
-            {
-                "a": [1, 2, 3, None, 2, 1, None],
-                "b": ["a", "z", "c", "a", "v", "z", "z"],
-            }
-        ),
-        lambda: cudf.DataFrame({"a": [], "b": []}),
-        lambda: cudf.DataFrame({"a": [None, None], "b": [None, None]}),
-        lambda: cudf.DataFrame(
-            {
-                "a": ["hello", "world", "rapids", "ai", "nvidia"],
-                "b": cudf.Series(
-                    [1, 21, 21, 11, 11],
-                    dtype="timedelta64[s]",
-                    index=["a", "b", "c", "d", " e"],
-                ),
-            },
-            index=["a", "b", "c", "d", " e"],
-        ),
-        lambda: cudf.DataFrame(
-            {
-                "a": ["hello", None, "world", "rapids", None, "ai", "nvidia"],
-                "b": cudf.Series(
-                    [1, 21, None, 11, None, 11, None], dtype="datetime64[s]"
-                ),
-            }
-        ),
-    ],
-)
-@pytest.mark.parametrize("numeric_only", [True, False])
-@pytest.mark.parametrize("dropna", [True, False])
-def test_dataframe_mode(df, numeric_only, dropna):
-    df = df()
-    pdf = df.to_pandas()
-
-    expected = pdf.mode(numeric_only=numeric_only, dropna=dropna)
-    actual = df.mode(numeric_only=numeric_only, dropna=dropna)
-    if len(actual.columns) == 0:
-        # pandas < 3.0 returns an Index[object] instead of RangeIndex
-        actual.columns = expected.columns
-    assert_eq(expected, actual, check_dtype=False)
-
-
-@pytest.mark.parametrize(
-    "lhs, rhs", [("a", "a"), ("a", "b"), (1, 1.0), (None, None), (None, "a")]
-)
-def test_equals_names(lhs, rhs):
-    lhs = cudf.DataFrame({lhs: [1, 2]})
-    rhs = cudf.DataFrame({rhs: [1, 2]})
-
-    got = lhs.equals(rhs)
-    expect = lhs.to_pandas().equals(rhs.to_pandas())
-
-    assert_eq(expect, got)
-
-
-def test_equals_dtypes():
-    lhs = cudf.DataFrame({"a": [1, 2.0]})
-    rhs = cudf.DataFrame({"a": [1, 2]})
-
-    got = lhs.equals(rhs)
-    expect = lhs.to_pandas().equals(rhs.to_pandas())
-
-    assert_eq(expect, got)
-
-
-@pytest.mark.parametrize(
-    "df1",
-    [
-        pd.DataFrame({"a": [10, 11, 12]}, index=["a", "b", "z"]),
-        pd.DataFrame({"z": ["a"]}),
-        pd.DataFrame({"a": [], "b": []}),
-    ],
-)
-@pytest.mark.parametrize(
-    "df2",
-    [
-        pd.DataFrame(),
-        pd.DataFrame({"a": ["a", "a", "c", "z", "A"], "z": [1, 2, 3, 4, 5]}),
-    ],
-)
-@pytest.mark.parametrize(
-    "op",
-    [
-        operator.eq,
-        operator.ne,
-        operator.lt,
-        operator.gt,
-        operator.le,
-        operator.ge,
-    ],
-)
-def test_dataframe_error_equality(df1, df2, op):
-    gdf1 = cudf.from_pandas(df1)
-    gdf2 = cudf.from_pandas(df2)
-
-    assert_exceptions_equal(op, op, ([df1, df2],), ([gdf1, gdf2],))
-
-
-@pytest.mark.parametrize(
-    "df,expected_pdf",
-    [
-        (
-            lambda: cudf.DataFrame(
-                {
-                    "a": cudf.Series([1, 2, None, 3], dtype="uint8"),
-                    "b": cudf.Series([23, None, None, 32], dtype="uint16"),
-                }
-            ),
-            pd.DataFrame(
-                {
-                    "a": pd.Series([1, 2, None, 3], dtype=pd.UInt8Dtype()),
-                    "b": pd.Series(
-                        [23, None, None, 32], dtype=pd.UInt16Dtype()
-                    ),
-                }
-            ),
-        ),
-        (
-            lambda: cudf.DataFrame(
-                {
-                    "a": cudf.Series([None, 123, None, 1], dtype="uint32"),
-                    "b": cudf.Series(
-                        [234, 2323, 23432, None, None, 224], dtype="uint64"
-                    ),
-                }
-            ),
-            pd.DataFrame(
-                {
-                    "a": pd.Series(
-                        [None, 123, None, 1], dtype=pd.UInt32Dtype()
-                    ),
-                    "b": pd.Series(
-                        [234, 2323, 23432, None, None, 224],
-                        dtype=pd.UInt64Dtype(),
-                    ),
-                }
-            ),
-        ),
-        (
-            lambda: cudf.DataFrame(
-                {
-                    "a": cudf.Series(
-                        [-10, 1, None, -1, None, 3], dtype="int8"
-                    ),
-                    "b": cudf.Series(
-                        [111, None, 222, None, 13], dtype="int16"
-                    ),
-                }
-            ),
-            pd.DataFrame(
-                {
-                    "a": pd.Series(
-                        [-10, 1, None, -1, None, 3], dtype=pd.Int8Dtype()
-                    ),
-                    "b": pd.Series(
-                        [111, None, 222, None, 13], dtype=pd.Int16Dtype()
-                    ),
-                }
-            ),
-        ),
-        (
-            lambda: cudf.DataFrame(
-                {
-                    "a": cudf.Series(
-                        [11, None, 22, 33, None, 2, None, 3], dtype="int32"
-                    ),
-                    "b": cudf.Series(
-                        [32431, None, None, 32322, 0, 10, -32324, None],
-                        dtype="int64",
-                    ),
-                }
-            ),
-            pd.DataFrame(
-                {
-                    "a": pd.Series(
-                        [11, None, 22, 33, None, 2, None, 3],
-                        dtype=pd.Int32Dtype(),
-                    ),
-                    "b": pd.Series(
-                        [32431, None, None, 32322, 0, 10, -32324, None],
-                        dtype=pd.Int64Dtype(),
-                    ),
-                }
-            ),
-        ),
-        (
-            lambda: cudf.DataFrame(
-                {
-                    "a": cudf.Series(
-                        [True, None, False, None, False, True, True, False],
-                        dtype="bool_",
-                    ),
-                    "b": cudf.Series(
-                        [
-                            "abc",
-                            "a",
-                            None,
-                            "hello world",
-                            "foo buzz",
-                            "",
-                            None,
-                            "rapids ai",
-                        ],
-                        dtype="object",
-                    ),
-                    "c": cudf.Series(
-                        [0.1, None, 0.2, None, 3, 4, 1000, None],
-                        dtype="float64",
-                    ),
-                }
-            ),
-            pd.DataFrame(
-                {
-                    "a": pd.Series(
-                        [True, None, False, None, False, True, True, False],
-                        dtype=pd.BooleanDtype(),
-                    ),
-                    "b": pd.Series(
-                        [
-                            "abc",
-                            "a",
-                            None,
-                            "hello world",
-                            "foo buzz",
-                            "",
-                            None,
-                            "rapids ai",
-                        ],
-                        dtype=pd.StringDtype(),
-                    ),
-                    "c": pd.Series(
-                        [0.1, None, 0.2, None, 3, 4, 1000, None],
-                        dtype=pd.Float64Dtype(),
-                    ),
-                }
-            ),
-        ),
-    ],
-)
-def test_dataframe_to_pandas_nullable_dtypes(df, expected_pdf):
-    actual_pdf = df().to_pandas(nullable=True)
-
-    assert_eq(actual_pdf, expected_pdf)
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        [{"a": 1, "b": 2, "c": 3}, {"a": 4, "b": 5, "c": 6}],
-        [{"a": 1, "b": 2, "c": None}, {"a": None, "b": 5, "c": 6}],
-        [{"a": 1, "b": 2}, {"a": 1, "b": 5, "c": 6}],
-        [{"a": 1, "b": 2}, {"b": 5, "c": 6}],
-        [{}, {"a": 1, "b": 5, "c": 6}],
-        [{"a": 1, "b": 2, "c": 3}, {"a": 4.5, "b": 5.5, "c": 6.5}],
-    ],
-)
-def test_dataframe_init_from_list_of_dicts(data):
-    expect = pd.DataFrame(data)
-    got = cudf.DataFrame(data)
-
-    assert_eq(expect, got)
-
-
-def test_dataframe_pipe():
-    pdf = pd.DataFrame()
-    gdf = cudf.DataFrame()
-
-    def add_int_col(df, column):
-        df[column] = df._constructor_sliced([10, 20, 30, 40])
-        return df
-
-    def add_str_col(df, column):
-        df[column] = df._constructor_sliced(["a", "b", "xyz", "ai"])
-        return df
-
-    expected = (
-        pdf.pipe(add_int_col, "one")
-        .pipe(add_int_col, column="two")
-        .pipe(add_str_col, "three")
-    )
-    actual = (
-        gdf.pipe(add_int_col, "one")
-        .pipe(add_int_col, column="two")
-        .pipe(add_str_col, "three")
-    )
-
-    assert_eq(expected, actual)
-
-    expected = (
-        pdf.pipe((add_str_col, "df"), column="one")
-        .pipe(add_str_col, column="two")
-        .pipe(add_int_col, "three")
-    )
-    actual = (
-        gdf.pipe((add_str_col, "df"), column="one")
-        .pipe(add_str_col, column="two")
-        .pipe(add_int_col, "three")
-    )
-
-    assert_eq(expected, actual)
-
-
-def test_dataframe_pipe_error():
-    pdf = pd.DataFrame()
-    gdf = cudf.DataFrame()
-
-    def custom_func(df, column):
-        df[column] = df._constructor_sliced([10, 20, 30, 40])
-        return df
-
-    assert_exceptions_equal(
-        lfunc=pdf.pipe,
-        rfunc=gdf.pipe,
-        lfunc_args_and_kwargs=([(custom_func, "columns")], {"columns": "d"}),
-        rfunc_args_and_kwargs=([(custom_func, "columns")], {"columns": "d"}),
-    )
-
-
-@pytest.mark.parametrize(
-    "op",
-    ["count", "kurt", "kurtosis", "skew"],
-)
-def test_dataframe_axis1_unsupported_ops(op):
-    df = cudf.DataFrame({"a": [1, 2, 3], "b": [8, 9, 10]})
-
-    with pytest.raises(
-        NotImplementedError, match="Only axis=0 is currently supported."
-    ):
-        getattr(df, op)(axis=1)
-
-
-def test_dataframe_from_pandas_duplicate_columns():
-    pdf = pd.DataFrame(columns=["a", "b", "c", "a"])
-    pdf["a"] = [1, 2, 3]
-
-    with pytest.raises(
-        ValueError, match="Duplicate column names are not allowed"
-    ):
-        cudf.from_pandas(pdf)
-
-
-@pytest.mark.parametrize(
-    "df",
-    [
-        pd.DataFrame(
-            {"a": [1, 2, 3], "b": [10, 11, 20], "c": ["a", "bcd", "xyz"]}
-        ),
-        pd.DataFrame(),
-    ],
-)
-@pytest.mark.parametrize(
-    "columns",
-    [
-        None,
-        ["a"],
-        ["c", "a"],
-        ["b", "a", "c"],
-        [],
-        pd.Index(["c", "a"]),
-        cudf.Index(["c", "a"]),
-        ["abc", "a"],
-        ["column_not_exists1", "column_not_exists2"],
-    ],
-)
-@pytest.mark.parametrize("index", [["abc", "def", "ghi"]])
-def test_dataframe_constructor_columns(df, columns, index, request):
-    def assert_local_eq(actual, df, expected, host_columns):
-        check_index_type = not expected.empty
-        if host_columns is not None and any(
-            col not in df.columns for col in host_columns
-        ):
-            assert_eq(
-                expected,
-                actual,
-                check_dtype=False,
-                check_index_type=check_index_type,
-            )
-        else:
-            assert_eq(
-                expected,
-                actual,
-                check_index_type=check_index_type,
-                check_column_type=False,
-            )
-
-    gdf = cudf.from_pandas(df)
-    host_columns = (
-        columns.to_pandas() if isinstance(columns, cudf.Index) else columns
-    )
-
-    expected = pd.DataFrame(df, columns=host_columns, index=index)
-    actual = cudf.DataFrame(gdf, columns=columns, index=index)
-
-    assert_local_eq(actual, df, expected, host_columns)
-
-
-def test_dataframe_constructor_column_index_only():
-    columns = ["a", "b", "c"]
-    index = ["r1", "r2", "r3"]
-
-    gdf = cudf.DataFrame(index=index, columns=columns)
-    assert not id(gdf["a"]._column) == id(gdf["b"]._column) and not id(
-        gdf["b"]._column
-    ) == id(gdf["c"]._column)
